+++ src/main/java/com/softuni/gameshop/model/DTO/GameDetailsDTO.java
@@ -2,6 +2,8 @@ package com.softuni.gameshop.model.DTO;
 
 import com.softuni.gameshop.model.enums.GenreNamesEnum;
 
+import java.math.BigDecimal;
+
 public class GameDetailsDTO {
 
     private Long id;
@@ -14,7 +16,7 @@ public class GameDetailsDTO {
 
     private Integer releaseYear;
 
-    private Double price;
+    private BigDecimal price;
 
     private String imageURL;
 
@@ -70,11 +72,11 @@ public class GameDetailsDTO {
         return this;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public GameDetailsDTO setPrice(Double price) {
+    public GameDetailsDTO setPrice(BigDecimal price) {
         this.price = price;
         return this;
     }
