+++ src/main/java/com/softuni/gameshop/service/ShoppingCartService.java
@@ -3,6 +3,7 @@ package com.softuni.gameshop.service;
 import com.softuni.gameshop.model.DTO.CartItemDTO;
 import com.softuni.gameshop.model.UserEntity;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ShoppingCartService {
@@ -14,5 +15,5 @@ public interface ShoppingCartService {
 
     UserEntity getCurrentUser();
 
-    double calculateTotalPrice();
+    BigDecimal calculateTotalPrice();
 }
